start of better context loading
