+++ README.md
@@ -7,9 +7,8 @@ dynamo - db integration for vigour-hub
 [![Coverage Status](https://coveralls.io/repos/github/vigour-io/blend-state-dynamo/badge.svg?branch=master)](https://coveralls.io/github/vigour-io/blend-state-dynamo?branch=master)
 
 ```javascript
-state.inject(require('blend-state-dynamo'))
-
 state.set({
+  inject: require('blend-state-dynamo')
   db: {
     id: AMAZON_ID,
     secret: AMAZON_SECRET,
@@ -21,8 +20,24 @@ state.set({
     }
   }
 })
+
 state.db.hasTable.is(true).then(() => {
-  t.ok(true, 'hasTable')
-  t.end()
+  console.log('got a table!')
 })
+```
+
+**Loading data**
+```javascript
+state.set({
+  inject: require('blend-state-dynamo'),
+  db: {
+    id: AMAZON_ID,
+    secret: AMAZON_SECRET,
+    table: testTable
+  }
+})
+
+state.db.load() // loads all data without context
+state.db.load('somecontextid') // loads all data with somecontextid
+state.db.load('*') //loads all data -- warn maybe very heavy
 ```
