+++ lib/api/load.js
@@ -5,7 +5,6 @@ module.exports = function (done) {
   const db = this
   const state = this.root
   // var lek // { exclusiveStartKey: lek } for large sets (latest ending key)
-
   db.hasTable.is(true, () => {
     db.db.scan({
       TableName: db.table.compute()
