+++ test/util.js
@@ -1,0 +1,13 @@
+'use strict'
+
+exports.out = (state) => {
+  const path = state.path()
+  const context = path[0]
+  path.shift()
+  return {
+    key: path,
+    context: context,
+    val: state.val,
+    stamp: state.stamp
+  }
+}
